Convert TesterUtils to extensions
